fix edit profile display, separate delete component
