+++ client/src/components/Layout/Navbar.js
@@ -9,6 +9,9 @@ import home from '../../image/home.svg';
 import homeActive from '../../image/home-active.svg';
 import findPeople from '../../image/find-people.svg';
 import findPeopleActive from '../../image/find-people-active.svg';
+//for logout
+import { connect } from 'react-redux';
+import { logoutUser } from '../../actions/authActions';
 
 class Navbar extends Component {
   constructor() {
@@ -83,4 +86,8 @@ class Navbar extends Component {
   }
 }
 
-export default Navbar;
+const mapStateToProps = (state) => ({
+  auth: state.auth,
+});
+
+export default connect(mapStateToProps, { logoutUser })(Navbar);
