+++ client/src/components/css/createprofile.css
@@ -1,0 +1,61 @@
+.create-profile {
+    font-size: 12px;
+    box-sizing: border-box;
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+} 
+
+.row {
+    width: 100%;
+    max-width: 935px;
+    margin: 30px auto 0;
+}
+
+.create-profile > .row {
+    display: flex;
+}
+
+.profile-info form{
+    background-color: #fff;
+    border: 2px solid #eee;
+    display: flex;
+    flex-direction: column;
+    padding: 15px 40px;
+    text-align: left;
+}
+
+.profile-info form input {
+    padding: 8px;
+    margin-bottom: 5px;
+    border: 2px solid #eee;
+    border-radius: 4px;
+
+}
+
+h3 {
+    margin-top: 20px;
+}
+
+.has-separator {
+    font-size: 20px;
+    color: #d11828;
+    text-transform: uppercase;
+    margin-top: 25px;
+    margin-bottom: 25px;
+    text-align: center;
+}
+
+@media(min-width: 767px) {
+    .sidebar-mobile {
+        display: none;
+    }
+}
+
+@media(max-width: 767px) {
+    .sidebar {
+        display: none;
+    }
+}
