--- client/src/components/css/editprofile.css
@@ -1,177 +1,0 @@
-.edit-profile {
-    font-size: 15px;
-    box-sizing: border-box;
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    padding-top: 100px;
-} 
-
-.profile-info {
-    height: 918px;
-  }
-
-  .row {
-    width: 100%;
-    max-width: 935px;
-    margin: 30px auto 0;
-}
-
-.edit-profile > .row {
-    display: flex;
-}
-
-.sidebar {
-    height: 200px;
-    flex: 1;
-}  
-
-.profile-info form {
-    background-color: #fff;
-    border: 2px solid #eee;
-    display: flex;
-    flex-direction: column;
-    padding: 30px 40px;
-    text-align: left;
-}
-
-h3 {
-    text-align: center;
-}
-
-p {
-    text-align: center;
-}
-
-.profile-info form input {
-    padding: 8px;
-    margin-bottom: 5px;
-    border: 2px solid #eee;
-    border-radius: 4px;
-
-}
-
-.btn {
-    margin: 10px 0 15px;
-    height: 30px;
-    background-color: #3897f6;
-    color: #fff,
-
-}
-
-
-.delete {
-    background-color: #d11828;
-
-}
-
-.has-separator {
-    font-size: 20px;
-    color: #d11828;
-    text-transform: uppercase;
-    margin-top: 25px;
-    margin-bottom: 25px;
-    text-align: center;
-}
-
-.form-data-omega form {
-    background-color: #fff;
-    border: 2px solid #eee;
-    display: flex;
-    flex-direction: column;
-    padding: 20px 40px;
-}
-
-.form-data-omega form input {
-    padding: 8px;
-    margin-bottom: 5px;
-    border: 2px solid #eee;
-    border-radius: 4px;
-
-}
-
-.card.upload-avatar {
-    margin:50px auto ;
-    width:300px;
-    border-radius: 25px; 
-    padding-bottom: 10px; 
-    background-color: rgba(var(--f23,255,255,255),1);
-    overflow: hidden;
-  }
-
-  div.card-header-av{
-    font-size:20px;
-    font-weight: 500;
-  }
- 
-  .userpost-avatar{
-    border-radius: 50%;
-    width: 110px;
-    height: 110px;  
-    vertical-align: middle;
-    justify-content: center;
-    margin-right: 10px;
-    margin-left: 25px;
-    margin-top: 5px;
-  }
-
-  .static-av {
-      border-radius: 50%;
-      width: 110px;
-      height: 110px;
-      justify-content: center;
-  }
-
-  .uploadav-form{
-    margin-top :20px;
-    justify-content: center;
-  }
-
-  .form-group {
-    width: 100%!important;
-  }
-  
-  div.upload-image{
-    text-align: center;
-  }
-  .fa.fa-file-image-o{
-    cursor: pointer;
-    font-size: 40px; 
-    color:#3897f0;
-  
-  }
-  .btn-primary{
-    background-color: transparent;
-    border-top: 1px solid rgba(var(--b6a,219,219,219),1);
-    margin: 0;
-    text-align: center;
-  }
-  .form-inline{
-    justify-content: center!important;
-  }
-  
-
-@media (max-width: 900px) {
-
-    .profile-info {
-        margin: 0 auto;
-    }
-
-    .edit-profile > .row {
-        padding: 0;
-    }
-}
-
-@media(min-width: 767px) {
-    .sidebar-mobile {
-        display: none;
-    }
-}
-
-@media (max-width: 767px) {
-    .sidebar {
-        display: none;
-    }
-}
