+++ package.json
@@ -25,6 +25,7 @@
     "mongoose": "^5.9.20",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
+    "react-toastify": "^6.0.8",
     "reactstrap": "^8.5.1",
     "validator": "^13.1.1"
   },
